coompleted tests and output to console

+++ ConferenceManagement/ConferenceManagement/Talk.cs
@@ -2,13 +2,13 @@
 {
     public class Talk
     {
-        public Talk(string topic, double duration)
+        public Talk(string topic, int duration)
         {
             Topic = topic;
             Duration = duration;
         }
 
         public string Topic { get; set; }
-        public double Duration { get; set; }
+        public int Duration { get; set; }
     }
 }
